requirements.txt


#Facial Fecognition
insightface
opencv-python
open_clip_torch
faiss-cpu
torchvision
onnxruntime

#OCR
ipywidgets
paddlepaddle
paddleocr
easyocr
pytesseract

#NLP
transformers
scikit-learn
torch
sentence-transformers
langchain
langchain-core
langchain-community
langchain-huggingface
accelerate
einops
pydantic
python-dotenv


# Other dependencies
numpy
pillow
tqdm
pathlib
matplotlib
pandas
ipywidgets
ultralytics
tensorflow


#audio
resemblyzer
speechbrain
librosa
soundfile
imageio-ffmpeg

#Database
chromadb